requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
charset-normalizer==2.0.7
environs==8.0.0
frozenlist==1.2.0
idna==3.3
marshmallow==3.14.1
multidict==5.2.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
urllib3==1.26.7
yarl==1.7.2
ado-py==0.0.7
aiodns==2.0.0
autopep8==1.5.4
cchardet==2.1.7
certifi>=2021.10.8
cffi==1.14.4
pycares==3.1.1
pycodestyle==2.6.0
pycparser==2.20
toml==0.10.2
typing-extensions==3.7.4.3
ujson==4.0.1
uvloop==0.14.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
asyncpg==0.22.0